refactor the code for perception
