+++ onboard/cv/meson.build
@@ -8,6 +8,8 @@ all_deps = [opencv, lcm]
 # TODO Jarvis will need to be smartened up to build this properly
 with_zed = get_option('with_zed')
 offline_test = get_option('offline_test')
+perception_debug = get_option('perception_debug')
+
 
 if with_zed
 	zed_include_dirs = include_directories('/usr/local/zed/include', '/usr/local/cuda/include')
@@ -36,12 +38,13 @@ endif
 conf_data = configuration_data()
 conf_data.set10('ZED_SDK_PRESENT', with_zed)
 conf_data.set10('OFFLINE_TEST', offline_test)
+conf_data.set10('PERCEPTION_DEBUG', perception_debug)
 configure_file(
 	input: 'config.h.in',
 	output: 'config.h',
 	configuration: conf_data)
 
 executable('onboard_cv',
-		   'main.cpp', 'camera.cpp',
+		   'main.cpp', 'camera.cpp', 'tennisball_detector.cpp', 'obstacle_detector.cpp',
 		   dependencies : all_deps,
 		   install : true)
